practice/app/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib[bcrypt]
pyjwt
pydantic[email]
jinja2
python-multipart